requirements.txt
numpy     
torch       
torchvision    
Pillow     
matplotlib  
seaborn 
albumentations
imgaug
tqdm
pyyaml        
jupyter           
notebook         
ipywidgets        
torchsummary
torchmetrics
pandas
scikit-learn